messaging_app/requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.13,<4.0
psycopg2-binary>=2.9,<3.0
python-decouple>=3.6,<4.0
django-cors-headers>=3.11,<4.0
drf-yasg>=1.20,<2.0
django-filter>=22.0,<23.0
djangorestframework-simplejwt>=5.2,<6.0
gunicorn>=20.1,<21.0
mysqlclient>=2.1,<2.2
drf-nested-routers>=0.93,<1.0  # ← CORRECTED PACKAGE NAME
